Move skill resource changes to generic tick event
