+++ src/clash/components.rs
@@ -208,6 +208,7 @@ pub fn create_world() -> World {
     ecs.subscribe(super::combat::field_event);
     ecs.subscribe(super::combat::explode_event);
     ecs.subscribe(super::defenses::defense_event);
+    ecs.subscribe(super::skills::tick_event);
 
     #[cfg(test)]
     {
