+++ src/clash/time.rs
@@ -1,10 +1,8 @@
-use std::cmp;
 use std::collections::BTreeMap;
 
-use ordered_float::*;
 use specs::prelude::*;
 
-use super::{SkillResourceComponent, TimeComponent};
+use super::{EventCoordinator, EventKind, TimeComponent};
 use crate::atlas::{EasyECS, EasyMutECS};
 
 pub const BASE_ACTION_COST: i32 = 100;
@@ -31,23 +29,18 @@ pub const EXHAUSTION_COST_PER_MOVE: f64 = 5.0;
 pub const FOCUS_PER_100_TICKS: f64 = 0.1;
 
 pub fn add_ticks(ecs: &mut World, ticks_to_add: i32) {
-    let mut times = ecs.write_storage::<TimeComponent>();
-    let mut skills = ecs.write_storage::<SkillResourceComponent>();
-    for (time, skill) in (&mut times, (&mut skills).maybe()).join() {
-        time.ticks += ticks_to_add;
-        if let Some(skill) = skill {
-            add_ticks_for_skill(skill, ticks_to_add);
+    let mut elements = vec![];
+    {
+        let entities = ecs.read_resource::<specs::world::EntitiesRes>();
+        let mut times = ecs.write_storage::<TimeComponent>();
+        for (entity, time) in (&entities, &mut times).join() {
+            time.ticks += ticks_to_add;
+            elements.push(entity);
         }
     }
-}
-
-fn add_ticks_for_skill(skill: &mut SkillResourceComponent, ticks_to_add: i32) {
-    let exhaustion_to_remove = EXHAUSTION_PER_100_TICKS as f64 * (ticks_to_add as f64 / 100.0);
-
-    let focus_to_add = FOCUS_PER_100_TICKS as f64 * (ticks_to_add as f64 / 100.0);
-    // Ordering f64 is hard _tm_
-    skill.exhaustion = *cmp::max(NotNan::new(0.0).unwrap(), NotNan::new(skill.exhaustion - exhaustion_to_remove).unwrap());
-    skill.focus = *cmp::min(NotNan::new(skill.max_focus).unwrap(), NotNan::new(skill.focus + focus_to_add).unwrap());
+    for e in elements {
+        ecs.raise_event(EventKind::Tick(ticks_to_add), Some(e));
+    }
 }
 
 pub fn wait_for_next(ecs: &mut World) -> Option<Entity> {
@@ -74,8 +67,6 @@ pub fn spend_time(ecs: &mut World, element: &Entity, ticks_to_spend: i32) {
 
 #[cfg(test)]
 mod tests {
-    use assert_approx_eq::assert_approx_eq;
-
     use super::super::{create_test_state, create_world, find_all_entities, find_at, find_at_time, find_first_entity, SkillResourceComponent};
     use super::*;
 
@@ -142,57 +133,5 @@ mod tests {
         let first = find_first_entity(&ecs);
         spend_time(&mut ecs, &first, BASE_ACTION_COST);
         assert_eq!(10, get_ticks(&ecs, &first));
-    }
-
-    #[test]
-    fn add_ticks_reduces_exhaustion() {
-        let mut ecs = create_test_state().with_character(2, 2, 100).with_map().build();
-        let player = find_at(&ecs, 2, 2);
-        ecs.write_storage::<SkillResourceComponent>().grab_mut(player).exhaustion = 50.0;
-        // This works as long as there is no rounding, as 20 *(5 *.5) = 50.0
-        for _ in 0..20 {
-            add_ticks(&mut ecs, 50);
-        }
-
-        {
-            let skills = ecs.read_storage::<SkillResourceComponent>();
-            assert_approx_eq!(skills.grab(player).exhaustion, 0.0);
-        }
-
-        // Keep going, make sure it doesn't drop below zero
-        for _ in 0..10 {
-            add_ticks(&mut ecs, 100);
-        }
-
-        {
-            let skills = ecs.read_storage::<SkillResourceComponent>();
-            assert_approx_eq!(skills.grab(player).exhaustion, 0.0);
-        }
-    }
-
-    #[test]
-    fn add_ticks_increases_focus() {
-        let mut ecs = create_test_state().with_character(2, 2, 100).with_map().build();
-        let player = find_at(&ecs, 2, 2);
-        ecs.write_storage::<SkillResourceComponent>().grab_mut(player).focus = 0.0;
-        ecs.write_storage::<SkillResourceComponent>().grab_mut(player).max_focus = 1.0;
-        for _ in 0..20 {
-            add_ticks(&mut ecs, 50);
-        }
-
-        {
-            let skills = ecs.read_storage::<SkillResourceComponent>();
-            assert_approx_eq!(skills.grab(player).focus, 1.0);
-        }
-
-        // Keep going, make sure it doesn't go above max
-        for _ in 0..10 {
-            add_ticks(&mut ecs, 100);
-        }
-
-        {
-            let skills = ecs.read_storage::<SkillResourceComponent>();
-            assert_approx_eq!(skills.grab(player).focus, 1.0);
-        }
     }
 }
